Upgraded intrinio to v3 framework
