+++ .changeset/chatty-otters-jump.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/intrinio-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
