+++ .changeset/cyan-apricots-smile.md
@@ -1,7 +1,6 @@
 ---
 '@chainlink/bank-frick-adapter': patch
 '@chainlink/iex-cloud-adapter': patch
-'@chainlink/intrinio-test-adapter': patch
 '@chainlink/kaiko-adapter': patch
 '@chainlink/nft-blue-chip-adapter': patch
 '@chainlink/openexchangerates-adapter': patch
