+++ packages/scripts/src/generate-readme/readmeBlacklist.json
@@ -38,7 +38,6 @@
     "graphql",
     "harmony",
     "historical-average",
-    "intrinio",
     "ipfs",
     "json-rpc",
     "kaiko",
