--- packages/sources/intrinio-test/CHANGELOG.md
@@ -1,118 +1,0 @@
-# @chainlink/intrinio-test-adapter
-
-## 1.3.3
-
-### Patch Changes
-
-- cc0d39064: Bumped framework version
-- fb1b11b77: Bumped framework version
-
-## 1.3.2
-
-### Patch Changes
-
-- 514171b76: Added rate limit tier for Chainlink Lab's plan
-
-## 1.3.1
-
-### Patch Changes
-
-- 590376288: Added accurate rate limit tiers
-
-## 1.3.0
-
-### Minor Changes
-
-- 11922dbf0: Removed WS_ENABLED env var
-
-## 1.2.2
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.2.1
-
-### Patch Changes
-
-- de8ce7352: Fixed websocket logic
-- 9bcb13c90: Bumped framework version
-
-## 1.2.0
-
-### Minor Changes
-
-- e0fe0a602: Added custom routing using WS_ENABLED env var
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
-
-## 1.1.7
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.1.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.1.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.1.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.2
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Removed WS_ENABLED env var. Upgraded typescript to version 4.9.5
-
-## 1.0.1
-
-### Patch Changes
-
-- e39eb6e6c: add missing stock endpoint alias
-
-## 1.0.0
-
-### Major Changes
-
-- cda6dc8bd: Added intrinio v3 EA
