--- packages/sources/intrinio-test/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@chainlink/intrinio-test-adapter",
-  "version": "1.3.3",
-  "description": "Chainlink intrinio-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "intrinio-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.29.12",
-    "axios": "^0.24.0",
-    "tslib": "2.4.0"
-  },
-  "devDependencies": {
-    "@sinonjs/fake-timers": "9.1.2",
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "@types/sinonjs__fake-timers": "8.1.2",
-    "nock": "13.2.9",
-    "typescript": "5.0.4"
-  }
-}
