--- packages/sources/intrinio-test/test-payload.json
@@ -1,5 +1,0 @@
-{
- "requests": [{
-  "base": "AAPL"
- }]
-}
