+++ packages/sources/intrinio/package.json
@@ -27,19 +27,17 @@
     "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
     "start": "yarn server:dist"
   },
+  "dependencies": {
+    "@chainlink/external-adapter-framework": "0.29.12",
+    "axios": "^0.24.0",
+    "tslib": "2.4.0"
+  },
   "devDependencies": {
+    "@sinonjs/fake-timers": "9.1.2",
     "@types/jest": "27.5.2",
     "@types/node": "16.11.51",
-    "@types/supertest": "2.0.12",
+    "@types/sinonjs__fake-timers": "8.1.2",
     "nock": "13.2.9",
-    "supertest": "6.2.4",
     "typescript": "5.0.4"
-  },
-  "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/ea-test-helpers": "workspace:*",
-    "axios": "^0.24.0",
-    "intrinio-realtime": "^2.3.0",
-    "tslib": "^2.3.1"
   }
 }
