--- packages/sources/intrinio/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/intrinio-adapter.json",
-  "title": "Chainlink External Adapter for Intrinio",
-  "description": "This adapter uses the Intrinio WS stream",
-  "required": ["API_KEY", "SYMBOLS"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "Your API client key"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
