--- packages/sources/intrinio/src/config/limits.json
@@ -1,13 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1s": 100
-    }
-  },
-  "ws": {
-    "default": {
-      "connections": 100,
-      "subscriptions": -1
-    }
-  }
-}
