+++ packages/sources/intrinio/src/index.ts
@@ -1,9 +1,29 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, makeWSHandler } from './adapter'
-import { makeConfig, NAME } from './config'
-import rateLimit from './config/limits.json'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { price } from './endpoint'
+import { config } from './config'
 
-const adapterContext = { name: NAME, rateLimit }
+export const adapter = new Adapter({
+  defaultEndpoint: price.name,
+  name: 'INTRINIO',
+  config,
+  endpoints: [price],
+  rateLimiting: {
+    tiers: {
+      bronze: {
+        rateLimit1m: 300,
+      },
+      chainlink: {
+        rateLimit1m: 600,
+      },
+      silver: {
+        rateLimit1m: 2000,
+      },
+      gold: {
+        rateLimit1m: 2000,
+      },
+    },
+  },
+})
 
-const { server } = expose(adapterContext, makeExecute(), makeWSHandler())
-export { NAME, makeExecute, makeConfig, server }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
