--- packages/sources/intrinio/types/intrinio-realtime/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'intrinio-realtime'
