+++ packages/tsconfig.test.json
@@ -381,9 +381,6 @@
       "path": "./sources/intrinio/tsconfig.test.json"
     },
     {
-      "path": "./sources/intrinio-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/ipfs/tsconfig.test.json"
     },
     {
